+++ src/Type/types.php
@@ -43,14 +43,11 @@ enum types implements Type
 
     /**
      * @no-named-arguments
-     * @psalm-suppress RedundantCondition
      * @param non-empty-string $class
      * @param non-empty-string $name
      */
     public static function alias(string $class, string $name, Type ...$arguments): Type
     {
-        \assert(array_is_list($arguments), 'Template arguments do not support names.');
-
         return new Internal\AliasType($class, $name, $arguments);
     }
 
@@ -339,7 +336,6 @@ enum types implements Type
 
     /**
      * @no-named-arguments
-     * @psalm-suppress RedundantCondition
      * @template TObject of object
      * @param class-string<TObject>|non-empty-string $class
      * @return ($class is class-string ? Type<TObject> : Type<object>)
@@ -349,8 +345,6 @@ enum types implements Type
         if ($class === \Closure::class && $arguments === []) {
             return self::closure;
         }
-
-        \assert(array_is_list($arguments), 'Template arguments do not support names.');
 
         return new Internal\NamedObjectType($class, $arguments);
     }
@@ -394,13 +388,10 @@ enum types implements Type
 
     /**
      * @no-named-arguments
-     * @psalm-suppress RedundantCondition
      * @param non-empty-string $name
      */
     public static function template(string $name, AtMethod|AtClass|AtFunction $declaredAt, Type ...$arguments): Type
     {
-        \assert(array_is_list($arguments), 'Template arguments do not support names.');
-
         return new Internal\TemplateType($name, $declaredAt, $arguments);
     }