+++ src/Type/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed `TypeVisitor::value()` and `DefaultVisitor::value()`.
 - Removed `TypeVisitor::classStringLiteral()` and `DefaultVisitor::classStringLiteral()`.
-- Added `@no-named-arguments` phpDoc and list assertion to `types::alias()`, `types::template()` and `types::object()`.
+- Added `@no-named-arguments` phpDocs to `types::alias()`, `types::template()` and `types::object()`.
 
 ## 0.3.2